Update Quantity
After adding the book specified in your cart, you can access your cart and update the quantity of this book as you like.
But do not forget any update of the quantity increase accompanied by an increase in the total price as well, and any update of the quantity decrease accompanied by a decrease in the total price as well.